+++ src/components/input/input.tsx
@@ -109,9 +109,9 @@ export class Input implements ComponentInterface, MutableAriaAttribute {
      *
      * @internal
      */
-    @Prop() fireFocusEvents = true;
+    @Prop() fireFocusEvents: boolean = true;
 
-    @State() private passwordReveal = false;
+    @State() private passwordReveal: boolean = false;
 
     /**
      * If the value of the type attribute is `"file"`, then this attribute will indicate the types of files that the
@@ -138,7 +138,7 @@ export class Input implements ComponentInterface, MutableAriaAttribute {
     /**
      * This Boolean attribute lets you specify that a form control should have input focus when the page loads.
      */
-    @Prop() autofocus = false;
+    @Prop() autofocus: boolean = false;
 
     /**
      * Set the amount of time, in milliseconds, to wait to trigger the `wcsInput` event after each keystroke.
@@ -164,7 +164,7 @@ export class Input implements ComponentInterface, MutableAriaAttribute {
     /**
      * If `true`, the user cannot interact with the input.
      */
-    @Prop({ reflect: true }) disabled = false;
+    @Prop({ reflect: true }) disabled: boolean = false;
 
     /**
      * A hint to the browser for which enter key to display.
@@ -212,7 +212,7 @@ export class Input implements ComponentInterface, MutableAriaAttribute {
      * If `true`, the user can enter more than one value. This attribute applies when the type attribute is set to
      * `"email"` or `"file"`, otherwise it is ignored.
      */
-    @Prop() multiple?: boolean;
+    @Prop({ reflect: true }) multiple?: boolean;
 
     /**
      * The name of the control, which is submitted with the form data.
@@ -240,17 +240,17 @@ export class Input implements ComponentInterface, MutableAriaAttribute {
     /**
      * If `true`, the user cannot modify the value.
      */
-    @Prop({ reflect: true }) readonly = false;
+    @Prop({ reflect: true }) readonly: boolean = false;
 
     /**
      * If `true`, the user must fill in a value before submitting a form.
      */
-    @Prop() required = false;
+    @Prop({ reflect: true }) required: boolean = false;
 
     /**
      * If `true`, the element will have its spelling and grammar checked.
      */
-    @Prop() spellcheck = false;
+    @Prop() spellcheck: boolean = false;
 
     /**
      * Specifies the state of the input. By default the input is in an normal state but you can to set it to 'error'